% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function}
\alias{cat_function}
\title{A Cat Function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\details{
Of all God’s creatures, there is only one that cannot be made slave of the lash. That one is the cat. If man could be crossed with the cat it would improve the man, but it would deteriorate the cat. - Mark Twain (1835-1910)
}
\examples{
cat_function(TRUE)
cat_function(FALSE)
}
\keyword{cats}
