% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_function.R
\name{house_function}
\alias{house_function}
\title{A Housing Function}
\usage{
house_function(live = TRUE)
}
\arguments{
\item{live}{Do you live in a house? Defaults to TRUE.}
}
\value{
The answer to your query
}
\description{
This function allows you to indicate your mode of residence.
}
\details{
It does not store anything and does precious little else than echo your housing status.
}
\examples{
house_function(TRUE)
house_function(FALSE)
}
\keyword{houses}
